+++ packages/frontend/src/features/projectGroupAccess/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectGroupAccessModal } from './components/ProjectGroupAccessModal';
